test(fault_manager): size a bag file without a cast that does nothing
std::filesystem::file_size and directory_entry::file_size already return
uintmax_t, which is size_t on every platform this builds for, so the cast
is identity and -Wuseless-cast flags it.

+++ src/ros2_medkit_fault_manager/test/test_fault_manager.cpp
@@ -1586,10 +1586,10 @@ struct ReportedBag {
 
     for (const auto & entry : std::filesystem::recursive_directory_iterator(dir)) {
       if (entry.is_regular_file()) {
-        footprint += static_cast<size_t>(entry.file_size());
+        footprint += entry.file_size();
       }
     }
-    served = static_cast<size_t>(std::filesystem::file_size(dir / storage_file));
+    served = std::filesystem::file_size(dir / storage_file);
   }
 
   ~ReportedBag() {
